fix estimated beats to print int instead of double

+++ src/WAVTrack.cpp
@@ -31,7 +31,7 @@ void WAVTrack::analyze_beatgrid() {
     // 3. Print number of beats and mention uncompressed precision
     // should print "  → Estimated beats: <beats>  → Precision factor: 1.0 (uncompressed audio)"
     double beats = (duration_seconds / 60.0) * bpm;
-    std::cout <<"  → Estimated beats: "<<beats<<"  → Precision factor: 1 (uncompressed audio)\n";
+    std::cout <<"  → Estimated beats: "<<(int)beats<<"  → Precision factor: 1 (uncompressed audio)\n";
 }
 
 double WAVTrack::get_quality_score() const {
